fix: fix raspberrypi firmware url
This fixes the url of the raspberry pi firmware.
In commit 15a683428278f8bf1b7c46110e6cb72d417ad1b8 this was changed from a
source github tarball to a release tarball. The release tarball does not
contain the DTB files needed.

Changing it to a source tarball of the fw release will probably fix
https://github.com/siderolabs/talos/issues/9981 (unable to crosscompile
u-boot on my machine)

Signed-off-by: Nico Berlee <nico.berlee@on2it.net>
Signed-off-by: Noel Georgi <git@frezbo.dev>